[TASK] Add PHP 8.0 and 8.1 to github actions
We need phpunit 8 to run PHP 8.0 and 8.1.

To stay compatible, we need to have a small hack
for setUp() and tearDown() - their signature must
have :void, which is not supported in low PHP versions.
We rename to custom methods and use @before / @after
annotation instead.

Another small hack is needed for phpunit exception
expectations.