[TASK] Fix PHP 8.4 compatibility
Defaulting to null without being nullable (?Foo) is deprecated for typed
parameters in PHP 8.4, therefore nullable is now specified explicitly.
This change requires at least PHP 7.1 (which has been made a minimum
with the parent commit).

Additional the call to str_getcsv is adapted to set all optional
arguments as the defaults will change in PHP >= 9 and PHP 8.4
deprecated the omission of the optional arguments.

Fixes #31